Render test app browser-side
It appears that server-side rendering doesn't play nice with npm link.
